% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCensus.R
\docType{data}
\name{dataCensus}
\alias{dataCensus}
\title{Census data}
\format{
A data frame with 23 variables.
}
\usage{
dataCensus
}
\description{
Census KDD Dataset (OpenML ID: 4535). Data frame with 50000 obs. of  23 variables, obtained via
\code{\link{getDataCensus}} with \code{target = "age"}.
}
\keyword{datasets}
