% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCensusFull.R
\docType{data}
\name{dataCensusFull}
\alias{dataCensusFull}
\title{Full census data set}
\format{
A data frame with 42 variables.
}
\usage{
dataCensusFull
}
\description{
Census KDD Dataset (OpenML ID: 4535). Data frame with 299285 obs. of  42 variables, obtained via
\code{OpenML::getOMLDataSet(data.id=4535, cache.only=cache.only)$data)}.
}
\keyword{datasets}
