% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funKerasGeneric.R
\name{evalKerasGeneric}
\alias{evalKerasGeneric}
\title{evalKerasGeneric model building and compile}
\usage{
evalKerasGeneric(x = NULL, kerasConf = NULL, specList = NULL)
}
\arguments{
\item{x}{matrix of transformed hyperparameter values to evaluate with the function.
If \code{NULL}, a simple keras model will be build, which is considered default
(see \code{\link{getSimpleKerasModel}}).}

\item{kerasConf}{List of additional parameters passed to keras as described in
\code{\link{getKerasConf}}. If no value is specified, stop() is called.}

\item{specList}{prepared data. See \code{\link{genericDataPrep}}.
See \code{\link{getGenericTrainValTestData}}.}
}
\value{
list with function values (training, validation, and test loss/accuracy,
and keras model information)
}
\description{
Hyperparameter Tuning: Keras Generic Classification Function.
}
\details{
Trains a simple deep NN on a generic data set.
Standard Code from \url{https://keras.rstudio.com/}.
Modified by T. Bartz-Beielstein.
}
\seealso{
\code{\link{getKerasConf}}

\code{\link{funKerasGeneric}}

\code{\link[keras]{fit}}
}
