% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{get_exp_config}
\alias{get_exp_config}
\title{Get experiment configuration}
\usage{
get_exp_config(task.type, model, task, rsmpl, data.seed)
}
\arguments{
\item{task.type}{character, either "classif" or "regr".}

\item{model}{character, learner chosen via mlr (here: "rpart", kknn", "cvglmnet","ranger","xgboost","svm).}

\item{task}{list, e.g., generated by \code{\link{get_task_census}}.}

\item{rsmpl}{resample strategy, e.g., generated by \code{makeFixedHoldoutInstance}.}

\item{data.seed}{this integer will be used to set the RNG before sampling the test and training data from the complete dataset.}
}
\value{
an mlr task with the respective data set
}
\description{
Get experiment configuration
}
