% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{get_objf}
\alias{get_objf}
\title{Get objective function}
\usage{
get_objf(config, timeout = 3600)
}
\arguments{
\item{config}{list}

\item{timeout}{integer, time in seconds after which a model (learner) evaluation will be aborted.}
}
\value{
an objective function that can be optimized via \code{\link[SPOT]{spot}}
}
\description{
mlrTools
This function receives a configuration for a tuning experiment,
and returns an objective function to be tuned via SPOT.
}
