% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{getBudget}
\alias{getBudget}
\title{Get budget (time, function evaluations)}
\usage{
getBudget(runtime_bench, settings)
}
\arguments{
\item{runtime_bench}{runtime}

\item{settings}{settings}
}
\value{
settings
}
\description{
Determine time factor and corrected  budget and timeout.
Runs a benchmark to determine processing speed of assigned computing resources.
}
\keyword{internal}
