% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentalDesign.R
\name{getExplan}
\alias{getExplan}
\title{Get experimental design}
\usage{
getExplan()
}
\value{
explan
}
\description{
Based on \code{\link{expand.grid}} an experimental design
is generated. Factors are: nfactors, nnumericals, cardinality, data.seed etc.
}
\keyword{internal}
