% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapParameters.R
\name{getIndices}
\alias{getIndices}
\title{Get indices (positions) of variable names}
\usage{
getIndices(model, a)
}
\arguments{
\item{model}{from \code{\link{getModelConf}}, e.g., \code{"dl"}.}

\item{a}{name of variables}
}
\value{
indices of variable names.
}
\description{
Get indices (positions) of variable names
}
\examples{
getIndices(model="dl",
           a = c("dropout", "units"))

}
