% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataKeras.R
\name{getMnistData}
\alias{getMnistData}
\title{getMnistData}
\usage{
getMnistData(kerasConf = getKerasConf())
}
\arguments{
\item{kerasConf}{List of additional parameters passed to keras as described in \code{\link{getKerasConf}}.
Default: \code{kerasConf = getKerasConf()}.}
}
\value{
list with training, validation, and test data
}
\description{
getMnistData
}
\examples{
\donttest{
### These examples require an activated Python environment as described in
### Bartz-Beielstein, T., Rehbach, F., Sen, A., and Zaefferer, M.:
### Surrogate Model Based Hyperparameter Tuning for Deep Learning with SPOT,
### June 2021. http://arxiv.org/abs/2105.14625.
PYTHON_RETICULATE = FALSE
if(PYTHON_RETICULATE){

library("SPOTMisc")
kerasConf <- getKerasConf()
mnist <- getMnistData()
}
}
}
\seealso{
\code{\link{getKerasConf}}

\code{\link{funKerasMnist}}
}
