% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{getObjf}
\alias{getObjf}
\title{Get objective function for mlr}
\usage{
getObjf(config, timeout = 3600)
}
\arguments{
\item{config}{list}

\item{timeout}{integer, time in seconds after which a model (learner) evaluation will be aborted.}
}
\value{
an objective function that can be optimized via \code{\link[SPOT]{spot}}.
}
\description{
mlrTools
This function receives a configuration for a tuning experiment,
and returns an objective function to be tuned via SPOT.
It basically provides the result from a call to \code{\link[mlr]{resample}}:
\code{resample(lrn, task, resample, measures = measures, show.info = FALSE)},
with measures defined as \code{mmce} for classification and \code{rmse} for regression, \code{timeboth}, \code{timetrain}, and
\code{timepredict}.
}
\details{
Parameter names are set, parameters are transformed
to the actual parameter scale, integer levels are converted
to factor levels for categorical parameters, and parameters
are set in relation to other parameters.
}
