% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{getPredf}
\alias{getPredf}
\title{Get predictions from mlr}
\usage{
getPredf(config, timeout = 3600)
}
\arguments{
\item{config}{list}

\item{timeout}{integer, time in seconds after which a model (learner) evaluation will be aborted.}
}
\value{
an prediction function that can be called via \code{\link[SPOT]{spot}}.
It basically provides the result from a call to \code{\link[mlr]{resample}}:
\code{resample(lrn, task, resample, measures = measures, show.info = FALSE)},
}
\description{
mlrTools
This function receives a configuration for a tuning experiment,
and returns predicted values.
}
