% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapParameters.R
\name{getVarNames}
\alias{getVarNames}
\title{Get variable names or subsets of variable names}
\usage{
getVarNames(model, i = "all")
}
\arguments{
\item{model}{from \code{\link{getModelConf}}, e.g., \code{"dl"}.}

\item{i}{index for selecting subsets. Default is \code{"all"}.}
}
\value{
vector of variable names. Returns \code{NA} if wrong indices are selected.
}
\description{
Get variable names or subsets of variable names
}
\examples{
# Default is return all:
getVarNames(model="dl")
getVarNames(model="dl",i=3)
getVarNames(model="dl",i=c(1,3,5))
# var name does not exits, so return NA
getVarNames(model="dl",i=c(100))


}
