% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotPlot.R
\name{ggplotProgress}
\alias{ggplotProgress}
\title{simple progress plot}
\usage{
ggplotProgress(
  dfRun,
  xlabel = "function evaluations",
  ylabel = "MMCE",
  aspectRatio = 2,
  scalesFreeFixed = "free_y",
  nColumns = 3
)
}
\arguments{
\item{dfRun}{data frame, e.g., result from \code{\link{prepareProgressPlot}}.}

\item{xlabel}{x label}

\item{ylabel}{y label}

\item{aspectRatio}{aspect.ratio}

\item{scalesFreeFixed}{"free_x", "free_y" or "fixed"}

\item{nColumns}{number of columns}
}
\value{
p ggplot
}
\description{
simple progress plot
}
