% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{int2fact}
\alias{int2fact}
\title{Helper function: transform integer to factor}
\usage{
int2fact(x, lvls)
}
\arguments{
\item{x}{an integer vector (that represents factor vector) to be transformed}

\item{lvls}{the original factor levels used}
}
\value{
the same factor, now coded with the original levels
}
\description{
This function re-codes a factor with pre-specified factor levels,
using an integer encoding as input.
}
