% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funKerasGeneric.R
\name{kerasBuildCompile}
\alias{kerasBuildCompile}
\title{evalKerasGeneric model building and compile}
\usage{
kerasBuildCompile(FLAGS, kerasConf, specList)
}
\arguments{
\item{FLAGS}{flags. list of hyperparameter values.
If \code{NULL}, a simple keras model will be build, which is considered default
(see \code{\link{getSimpleKerasModel}}).}

\item{kerasConf}{List of additional parameters passed to keras as described in \code{\link{getKerasConf}}.
Default: \code{kerasConf = getKerasConf()}.}

\item{specList}{prepared data. See \code{\link{genericDataPrep}}.
See \code{\link{getGenericTrainValTestData}}.}
}
\value{
list with function values (training, validation, and test loss/accuracy,
and keras model information)
}
\description{
Hyperparameter Tuning: Keras Generic Classification Function.
}
\details{
Trains a simple deep NN on a generic data set.
Standard Code from \url{https://tensorflow.rstudio.com/}
Modified by T. Bartz-Beielstein (tbb@bartzundbartz.de)
}
\seealso{
\code{\link{getKerasConf}}

\code{\link{funKerasGeneric}}

\code{\link[keras]{fit}}
}
