% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funKerasGeneric.R
\name{kerasReturnDummy}
\alias{kerasReturnDummy}
\title{Return dummy values}
\usage{
kerasReturnDummy(kerasConf)
}
\arguments{
\item{kerasConf}{keras configuration list}
}
\value{
y row matrix of random (uniformly distributed) return values
}
\description{
Return dummy values
}
\examples{
kerasConf <- getKerasConf()
kerasReturnDummy(kerasConf)

}
