% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{makeBench}
\alias{makeBench}
\title{Make benchmark}
\usage{
makeBench()
}
\value{
total time (in seconds)
}
\description{
Run standard benchmark from the package \code{benchmarkme} and return the
required time in seconds
#'
}
\keyword{internal}
