% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{makeLearnerFromHyperparameters}
\alias{makeLearnerFromHyperparameters}
\title{Make mlr learner from conf and hyperparameter vector}
\usage{
makeLearnerFromHyperparameters(x = NULL, cfg = NULL)
}
\arguments{
\item{x}{hyperparameter vector}

\item{cfg}{configuration list}
}
\value{
mlr learner
}
\description{
calls makelearner
}
