% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{plot_pareto}
\alias{plot_pareto}
\title{Pareto front (as well as non-optimal solutions)}
\usage{
plot_pareto(object, xvar = 1, yvar = 2, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{object}{the result list returned by \code{\link[SPOT]{spot}}, importantly including a \code{modelFit}, and the data \code{x}, \code{y}.}

\item{xvar}{integer which specifies the variable that is displayed on the x axis. xvar=1 (default) means that the y-variable is shown (tuned measure). Larger integers mean that respective columns from logInfo are used (then, ylog specifies the respective column number, starting with 2 for the first logged value).}

\item{yvar}{integer which specifies the variable that is displayed on the color scale. yvar==1 (default) means that the y-variable is shown (tuned measure). Larger integers mean that respective columns from logInfo are used (then, ylog specifies the respective column number, starting with 2 for the first logged value).}

\item{xlab}{character, giving the label for the x-axis of the plot.}

\item{ylab}{character, giving the label for the y-axis of the plot.}

\item{...}{additional parameters (currently unused).

By default, it is assumed that all variables that are plotted (yvar,xvar) are minimized.
If this is not true, assign a negative sign to the respective integer of the xvar and yvar arguments.}
}
\value{
plotly visualization (type 'scatter'), based on \code{\link[plotly]{plot_ly}}
}
\description{
mlrTools
}
\seealso{
\code{\link[SPOT]{plotFunction}}, \code{\link[SPOT]{plotData}}
}
