% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNice.R
\name{plotnice.spotTreeModel}
\alias{plotnice.spotTreeModel}
\title{Plot a nice rpart tree model}
\usage{
plotnice.spotTreeModel(x, ...)
}
\arguments{
\item{x}{tree model (settings and parameters) of class \code{spotTreeModel}.}

\item{...}{parameters passed to rpart.plot plotting function}
}
\value{
The returned value is identical to that of \code{\link[rpart.plot]{prp}}.
}
\description{
Plot model produced by \code{\link[SPOT]{buildTreeModel}}.
}
\keyword{internal}
