% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMarkovChain.R
\name{plotRegionByName}
\alias{plotRegionByName}
\title{plotRegionByName}
\usage{
plotRegionByName(regionList, country = "Germany")
}
\arguments{
\item{regionList}{A list containing a representation of the data.}

\item{country}{Name of a country from the list dataList}
}
\value{
A plot
}
\description{
Plot confirmed cases and fatalities (cumulative) for one country/region by region name.
}
\details{
Data are taken from the \code{regionTrain} and \code{regionPopulation} data sets
that were combined using the \code{\link{preprocessInputData}} function.
\code{regionTrain} and \code{regionPopulation} are stored in the
\code{babsim.data} package.
}
\examples{
regionList <- preprocessInputData(regionTrain, regionPopulation)
p <- plotRegionByName(regionList = regionList, country = "Germany")
}
\seealso{
\code{\link{plotRegion}}.
}
