% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{prepare_spot_result_plot}
\alias{prepare_spot_result_plot}
\title{Prepare data (results from a tuning run) for plots}
\usage{
prepare_spot_result_plot(data, model = buildRanger, modelControl = list(), ...)
}
\arguments{
\item{data}{a list containing the various data, e.g., as produced by a \code{\link[SPOT]{spot}} call.}

\item{model}{a function that can be used to build a model based on the data, e.g. : \code{buildRanger} or \code{buildKriging}. Default is \code{buildRanger}, since it is fast and robust.}

\item{modelControl}{a list of control settings for the respective model. Default is an empty list (use default model controls).}

\item{...}{additional parameters passed to \code{\link{prepare_data_plot}}: \code{namesx}, \code{namesy}, \code{nameslog} character vectors providing names for x, y and logInfo data. Se}
}
\value{
list with plotting data and information generated with \code{\link{prepare_data_plot}}
}
\description{
Preparation ot the list elements used for plotting.
}
