% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMarkovChain.R
\name{preprocessCdeInputData}
\alias{preprocessCdeInputData}
\title{preprocessCdeInputData}
\usage{
preprocessCdeInputData(cdeData)
}
\arguments{
\item{cdeData}{data frame}
}
\value{
A large list containing the following information (3 lists) for each region:
\describe{
     \item{\code{date}}{Date}
     \item{\code{confirmed}}{num Number of confirmed cases (cumulative)}
     \item{\code{fatalities}}{num Number of fatalities (cumulative)}
}
}
\description{
Prepare Ced Data for SIR modeling
}
\details{
Resulting data set can be used as input for the COVID-19 simulations of the
\code{\link{tuneRegionModel}} function.
}
