% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerasHelper.R
\name{printf}
\alias{printf}
\title{formatted output}
\usage{
printf(...)
}
\arguments{
\item{...}{output to be printed}
}
\description{
Combine \code{\link{sprintf}} and \code{\link{writeLines}} to
generate formatted output
}
\examples{

x <- 123
printf("x value: \%d", x)

}
