% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funKerasModular.R
\name{printFLAGS}
\alias{printFLAGS}
\title{Print parameter values from FLAG list}
\usage{
printFLAGS(FLAGS)
}
\arguments{
\item{FLAGS}{list of parameters, see \code{\link{mapX2FLAGS}}}
}
\description{
Print values of the following hyperparmeters:
\describe{
	\item{\code{x[1]: dropout}}{dropout rate first layer.}
	\item{\code{x[2]: dropoutfac}}{dropout factor (multiplier).}
	\item{\code{x[3]: units}}{number of units in the first layer.}
	\item{\code{x[4]: unitsfact}}{units factor (multiplier).}
	\item{\code{x[5]: learning_rate}}{learning rate for optimizer. See, e.g.: \code{link{optimizer_sgd}}}
	\item{\code{x[6]: epochs}}{number of training epochs.}
	\item{\code{x[7]: beta_1}}{The exponential decay rate for the 1st moment estimates. float, 0 < beta < 1. Generally close to 1.}
	\item{\code{x[8]: beta_2}}{The exponential decay rate for the 2nd moment estimates. float, 0 < beta < 1. Generally close to 1.}
	\item{\code{x[9]: layers}}{number of layers.}
 	\item{\code{x[10]: epsilon}}{float >= 0. Fuzz factor. If NULL, defaults to k_epsilon().}
 	\item{\code{x[11]: name}}{integer. Specifies optimizer.}
}
}
