% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionPopulation.R
\docType{data}
\name{regionPopulation}
\alias{regionPopulation}
\title{Region Population Data}
\format{
A data frame with 313 rows and 2 columns:
\describe{
  \item{regionName}{chr: Country.  ("Afghanistan/"--""Zimbabwe/")}
  \item{population}{int: Population size. (825--1366417754)}
}
}
\source{
<http://owos.gm.fh-koeln.de:8055/bartz/bartz-data/blob/cd4aad9a6640f7c36caa53c898a78568e34374e8/Covid-19/regionPopulation.csv>
}
\usage{
regionPopulation
}
\description{
A data set of populations.
}
\seealso{
\code{\link{regionTrain}} and \code{\link{regionTest}}.

\code{cde20200813} and \code{DEcde20200813}, which are part of the 
\code{babsim.data} package, because these are relatively large data sets.
}
\keyword{datasets}
