% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resKerasMnist03.R
\docType{data}
\name{resKerasMnist03}
\alias{resKerasMnist03}
\title{Results from a very simple SPOT run, version 3, with replicates}
\format{
A list of 9 values
\describe{
  \item{xbest}{num:}
  \item{ybest}{num:}
  \item{x}{num}
  \item{y}{num}
  \item{logInfo}{logi}
  \item{count}{int}
  \item{msg}{chr}
  \item{model fit}{List of 13}
  \item{ybestVec}{num}
 }
}
\usage{
resKerasMnist03
}
\description{
Details and the corresponding R code to generate the data
can be found in the package vignette \code{SPOTMiscVignette.Rmd}.
}
\keyword{datasets}
