% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resKerasTransferLearning04.R
\docType{data}
\name{resKerasTransferLearning04}
\alias{resKerasTransferLearning04}
\title{Results from a SPOT run, version 4, with replicates}
\format{
A list of 9 values
\describe{
  \item{xbest}{num:}
  \item{ybest}{num:}
  \item{x}{num}
  \item{y}{num}
  \item{logInfo}{logi}
  \item{count}{int}
  \item{msg}{chr}
  \item{model fit}{List of 13}
  \item{ybestVec}{num}
 }
}
\usage{
resKerasTransferLearning04
}
\description{
SPOT results: resKerasTransferLearning04

Details and the corresponding R code to generate the data
can be found in the package vignette \code{SPOTMiscVignette.Rmd}.
}
\keyword{datasets}
