% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resKerasTransferLearning05.R
\docType{data}
\name{resKerasTransferLearning05}
\alias{resKerasTransferLearning05}
\title{Results from a SPOT run, version 5, with replicates: negValidationAccuracy}
\format{
A list of 9 values
\describe{
  \item{xbest}{num:}
  \item{ybest}{num:}
  \item{x}{num}
  \item{y}{num}
  \item{logInfo}{logi}
  \item{count}{int}
  \item{msg}{chr}
  \item{model fit}{List of 13}
  \item{ybestVec}{num}
 }
}
\usage{
resKerasTransferLearning05
}
\description{
SPOT results: resKerasTransferLearning05

Details and the corresponding R code to generate the data
can be found in the package vignette \code{SPOTMiscVignette.Rmd}.
}
\keyword{datasets}
