% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resKerasTransferLearning06.R
\docType{data}
\name{resKerasTransferLearning06}
\alias{resKerasTransferLearning06}
\title{Results from a SPOT run, version 06, default}
\format{
A list of 9 values
\describe{
  \item{xbest}{num:}
  \item{ybest}{num:}
  \item{x}{num}
  \item{y}{num}
  \item{logInfo}{logi}
  \item{count}{int}
  \item{msg}{chr}
  \item{model fit}{List of 13}
  \item{ybestVec}{num}
 }
}
\usage{
resKerasTransferLearning06
}
\description{
SPOT results: resKerasTransferLearning06: Evaluation of the default
parameters

Details and the corresponding R code to generate the data
can be found in the package vignette \code{SPOTMiscVignette.Rmd}.
}
\keyword{datasets}
