% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotStats.R
\name{RMSE}
\alias{RMSE}
\title{root mean squared errors}
\usage{
RMSE(y, yhat)
}
\arguments{
\item{y}{actual value}

\item{yhat}{predicted value}
}
\value{
root mean squared errors
}
\description{
root mean squared errors
}
