% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotStats.R
\name{scorePredictions}
\alias{scorePredictions}
\title{Score results from pred}
\usage{
scorePredictions(val)
}
\arguments{
\item{val}{list of matrices with true and
predicted values, e.g., output from
\code{\link{predMlCensus}}}
}
\value{
matrix with scores
}
\description{
errors for \code{(actual, predicted)} values.
Based on package \code{Metrics}.
}
