% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerasActivation.R
\name{selectKerasActivation}
\alias{selectKerasActivation}
\title{Select keras activation function}
\usage{
selectKerasActivation(activation)
}
\arguments{
\item{activation}{integer specifying the activation function. Can be one of the following:
\code{1=NULL}, \code{2=RELU}}
}
\value{
activation function use with \code{\link{funKerasMnist}}.
}
\description{
Select keras activation function
}
