% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotStats.R
\name{SSE}
\alias{SSE}
\title{sum of squared errors}
\usage{
SSE(y, yhat)
}
\arguments{
\item{y}{actual value}

\item{yhat}{predicted value}
}
\value{
sum of squared errors
}
\description{
sum of squared errors
}
