% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{trans_10pow}
\alias{trans_10pow}
\title{10 power x transformation}
\usage{
trans_10pow(x)
}
\arguments{
\item{x}{input}
}
\value{
\code{10^x}
}
\description{
Parameter values can be translated,
e.g., to base 10.
}
