% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{trans_id}
\alias{trans_id}
\title{Identity transformation}
\usage{
trans_id(x)
}
\arguments{
\item{x}{input}
}
\value{
\code{x}
}
\description{
Parameter values can be translated,
e.g., to base 10 as implemented in \code{\link{trans_10pow}}.
\code{trans_id} implements the identity (transformation), i.e., x is mapped to x.
}
