% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotTransform.R
\name{trans_mult2_round}
\alias{trans_mult2_round}
\title{Mult 2 transformation}
\usage{
trans_mult2_round(x)
}
\arguments{
\item{x}{input}
}
\value{
\code{x}
}
\description{
Parameter values can be translated,
implements the multiplication (transformation), i.e., x is mapped
to round(2x).
}
