% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotTransform.R
\name{trans_odd_round}
\alias{trans_odd_round}
\title{odd transformation}
\usage{
trans_odd_round(x)
}
\arguments{
\item{x}{input}
}
\value{
\code{x}
}
\description{
Generate odd numbers, i.e., \code{x -> 2x-1} for x > 0.
Return values are rounded using \code{round}.
}
\examples{
f2 <- function(x){2^x}
fn <- c("trans_odd_round", "exp", "f2")
xNat <- diag(3)
SPOT::transformX(xNat, fn)

}
