% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotTransform.R
\name{translate_levels}
\alias{translate_levels}
\title{Helper function: translate levels}
\usage{
translate_levels(x, translations)
}
\arguments{
\item{x}{a factor vector to be translated}

\item{translations}{a named list that specifies the translation: \code{list(newlevel=c(oldlevel1,oldlevel2,etc))}.}
}
\value{
translated factor
}
\description{
Translate existing levels of a factor into new levels.
}
