% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTools.R
\name{valid_inputs}
\alias{valid_inputs}
\title{Check the validity of input parameters.}
\usage{
valid_inputs(chars)
}
\arguments{
\item{chars}{character}
}
\value{
correct characters
}
\description{
Helper function. Check correct parameter names.
}
\keyword{internal}
