% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerasHelper.R
\name{genCatsDogsData}
\alias{genCatsDogsData}
\title{generate Cats Dogs Data}
\usage{
genCatsDogsData(kerasConf = getKerasConf())
}
\arguments{
\item{kerasConf}{keras configuration. Default: \code{kerasConf = \link{getKerasConf}}}
}
\value{
list with test, validation, and test data
}
\description{
Generate data for \code{\link{funKerasTransferLearning}}
}
\details{
Standard Data from https://keras.rstudio.com/
Modified by T. Bartz-Beielstein (tbb@bartzundbartz.de)
}
