% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapParameters.R
\name{printFLAGS}
\alias{printFLAGS}
\title{Print parameter values from FLAG list}
\usage{
printFLAGS(FLAGS)
}
\arguments{
\item{FLAGS}{list of parameters, see \code{\link{mapX2FLAGS}}}
}
\description{
Simple print method for FLAG list.
}
