% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotTransform.R
\name{trans_2pow_round}
\alias{trans_2pow_round}
\title{2 power x transformation with round}
\usage{
trans_2pow_round(x)
}
\arguments{
\item{x}{input}
}
\value{
\code{round(2^x)}
}
\description{
Parameter values can be translated,
e.g., to base 10 as implemented in \code{\link{trans_10pow}}.
\code{trans_2pow_round} implements the transformation x -> round(2^x).
}
