% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotTransform.R
\name{trans_id}
\alias{trans_id}
\title{Identity transformation}
\usage{
trans_id(x)
}
\arguments{
\item{x}{input}
}
\value{
\code{x}
}
\description{
Parameter values can be translated,
e.g., to base 10 as implemented in \code{\link{trans_10pow}}.
\code{trans_id} implements the identity (transformation), i.e., x is mapped to x.
}
\examples{
f2 <- function(x){2^x}
fn <- c("identity", "exp", "f2")
xNat <- diag(3)
SPOT::transformX(xNat, fn)

}
